Struct riddle_font::ImgFont

pub struct ImgFont { /* fields omitted */ }

Represents an image font, which is an Image containing glyphs, glyph information specifying which parts of the image map to which glyph, and layout information.

Implementations

impl ImgFont

pub fn vertical_spacing(&self) -> u32

This represents the distance, in pixels, between the top of one line of text and the next line of text.

Example

let ttf_font = TTFont::load(&font_bytes[..])?;
let img_font = ImgFontGenerator::new("ABCDEF", 32).generate(&ttf_font)?;

assert!(img_font.vertical_spacing() > 0);

pub fn image(&self) -> &Image

The image storing the glyph image data.

Example

let ttf_font = TTFont::load(&font_bytes[..])?;
let img_font = ImgFontGenerator::new("ABCDEF", 32).generate(&ttf_font)?;

assert!(img_font.image().width() > 0);

pub fn glyphs(&self) -> &HashMap<char, ImgFontGlyph>

The mapping from char to glyphs registered with the font.

Example

let ttf_font = TTFont::load(&font_bytes[..])?;
let img_font = ImgFontGenerator::new("ABCDEF", 32).generate(&ttf_font)?;

assert_eq!(6, img_font.glyphs().len());
assert_eq!('A', img_font.glyphs().get(&'A').unwrap().character);

pub fn render_simple(&self, text: &str) -> Result<Image, FontError>

Render a string to a riddle_image::Image using the font at its native scale.

Example

let ttf_font = TTFont::load(&font_bytes[..])?;
let img_font = ImgFontGenerator::new("ABCDEF", 32).generate(&ttf_font)?;

let rendered_img = img_font.render_simple("AABCA")?;

assert_eq!(rendered_img.height(), img_font.vertical_spacing());
assert!(rendered_img.width() > 0);

pub fn layout<F: FnMut(char, &Rect<u32>, Vector2<u32>)>(&self, text: &str, f: F)

Layout a string using the font, relative to a (0,0) point representing the top left of the layout space.

The callback fuction is invoked once for each character, providing the part of the image to be used, and the location to draw that image.

Example

let ttf_font = TTFont::load(&font_bytes[..])?;
let img_font = ImgFontGenerator::new("ABCDEF", 32).generate(&ttf_font)?;

let mut last_x = 0;
let mut total_width = 0;
let mut string = String::from("");

img_font.layout("AABCA", |c, rect, location| {
    assert!(location.x >= last_x);
    last_x = location.x;
    total_width += rect.dimensions.x;
    string += c.to_string().as_str();
});

assert!(last_x > 0);
assert!(total_width > 0);
assert_eq!("AABCA", string);